Fixing signedness in gio/gio-launch-desktop.c
io/gio-launch-desktop.c: In function ‘main’:
gio/gio-launch-desktop.c:46:9: error: comparison of integer expressions of different signedness: ‘int’ and ‘long unsigned int’ [-Werror=sign-compare]
   if (r >= sizeof (buf))
         ^~

+++ gio/gio-launch-desktop.c
@@ -37,7 +37,7 @@ main (int argc, char *argv[])
 {
   pid_t pid = getpid ();
   char buf[50];
-  int r;
+  size_t r;
 
   if (argc < 2)
     return -1;
